Added separate functionality and storage for vocal tract videos and vowel sound videos.
This also had the effect of changing the way recordings are stored in
the database - they now have an optional vocaltract singlefile field.
The category in the file name is now used to differentiate between
videos and vocal tracts.
The combobox in the video player now holds recording objects, not word
objects,.